requirements.txt
flask>=2.0.0
line-bot-sdk>=2.4.2
openai>=1.0.0
python-dotenv>=0.19.0
tenacity>=8.0.0
requests
pydub
ffmpeg-python